+++ src/core/configuration/config_value.h
