+++ src/core/configuration/configuration_manager.cpp
@@ -1,4 +1,4 @@
-#include "ConfigurationManager.h"
+#include "configuration_manager.h"
 #include "logging.h"
 #include <algorithm>
 #include <set>
