+++ src/core/configuration/iconfiguration_change_listener.h
